+++ src/HTS.Data/Entity/ProformaProcess.cs
@@ -23,7 +23,7 @@ namespace HTS.Data.Entity
         [Required]
         public decimal ProformaPrice { get; set; }
         [Required]
-        public int Change { get; set; }
+        public decimal Change { get; set; }
         [Required]
         public decimal ProformaFinalPrice { get; set; }
